Move feedback form components to settings tab
