+++ components/settings/QRScanner.js
